+++ modules/home/profiles/developer/default.nix
@@ -4,10 +4,14 @@
     ../../features/starship
     ../../features/aws
     ../../features/developer-tools
+    ../../features/noctalia
+    ../../features/niri
   ];
 
   homeFeatures.shell.enable = true;
   homeFeatures.starship.enable = true;
   homeFeatures.aws.enable = true;
   homeFeatures.developerTools.enable = true;
+  homeFeatures.noctalia.enable = true;
+  homeFeatures.niri.enable = true;
 }