+++ modules/nixos/profiles/niri-desktop/default.nix
@@ -1,0 +1,12 @@
+{ ... }:
+
+{
+  imports = [
+    ../../features/desktop
+  ];
+
+  features.desktop = {
+    enable = true;
+    environment = "niri";
+  };
+}